Update WindowsServiceLifecycle.java
Changed from using the Windows service wrappers "restart" command to using "restart!". This resolves the issue of the service wrapper process killing itself as part of the non-atomic stop-wait-start handling of restart.

See JENKINS-22685 for further description of the issue, and see WinSW README "Restarting service from itself" for info on the "restart!" command.